feat(explorer): implement useFormattedDate hook for date formatting and tooltip display
