+++ apps/explorer/src/hooks/index.ts
@@ -18,3 +18,4 @@ export * from './useNormalizedMoveModule';
 export * from './useSearch';
 export * from './useVerifiedSourceCode';
 export * from './useEndOfEpochTransactionFromCheckpoint';
+export * from './useFormattedDate';
